Deduplicate existing stubs (a bit)
