+++ cni/pkg/iptables/iptables_test.go
@@ -15,13 +15,14 @@
 package iptables
 
 import (
-	"io"
 	"net/netip"
 	"path/filepath"
 	"strings"
 	"testing"
 
 	testutil "istio.io/istio/pilot/test/util"
+
+	dep "istio.io/istio/tools/istio-iptables/pkg/dependencies"
 )
 
 func TestIptables(t *testing.T) {
@@ -45,7 +46,7 @@ func TestIptables(t *testing.T) {
 				cfg := constructTestConfig()
 				cfg.EnableInboundIPv6 = ipv6
 				tt.config(cfg)
-				ext := &DependenciesStub{}
+				ext := &dep.DependenciesStub{}
 				iptConfigurator := NewIptablesConfigurator(cfg, ext, EmptyNlDeps())
 				var probeIP *netip.Addr
 				if ipv6 {
@@ -84,7 +85,7 @@ func TestIptablesHostRules(t *testing.T) {
 				cfg := constructTestConfig()
 				cfg.EnableInboundIPv6 = ipv6
 				tt.config(cfg)
-				ext := &DependenciesStub{}
+				ext := &dep.DependenciesStub{}
 				iptConfigurator := NewIptablesConfigurator(cfg, ext, EmptyNlDeps())
 				var probeIP *netip.Addr
 				if ipv6 {
@@ -118,7 +119,7 @@ func TestInvokedTwiceIsIdempotent(t *testing.T) {
 
 	cfg := constructTestConfig()
 	tt.config(cfg)
-	ext := &DependenciesStub{}
+	ext := &dep.DependenciesStub{}
 	iptConfigurator := NewIptablesConfigurator(cfg, ext, EmptyNlDeps())
 	err := iptConfigurator.CreateInpodRules(&probeSNATipv4)
 	if err != nil {
@@ -126,7 +127,7 @@ func TestInvokedTwiceIsIdempotent(t *testing.T) {
 	}
 	compareToGolden(t, false, tt.name, ext.ExecutedAll)
 
-	*ext = DependenciesStub{}
+	*ext = dep.DependenciesStub{}
 	// run another time to make sure we are idempotent
 	err = iptConfigurator.CreateInpodRules(&probeSNATipv4)
 	if err != nil {
@@ -156,41 +157,5 @@ func compareToGolden(t *testing.T, ipv6 bool, name string, actual []string) {
 func constructTestConfig() *Config {
 	return &Config{
 		RestoreFormat: false,
-	}
-}
-
-type DependenciesStub struct {
-	ExecutedNormally []string
-	ExecutedQuietly  []string
-	ExecutedAll      []string
-	Stdin            []byte
-}
-
-func (s *DependenciesStub) RunOrFail(cmd string, stdin io.ReadSeeker, args ...string) {
-	if stdin != nil {
-		s.Stdin, _ = io.ReadAll(stdin)
-	}
-	s.execute(false /*quietly*/, cmd, args...)
-}
-
-func (s *DependenciesStub) Run(cmd string, stdin io.ReadSeeker, args ...string) error {
-	if stdin != nil {
-		s.Stdin, _ = io.ReadAll(stdin)
-	}
-	s.execute(false /*quietly*/, cmd, args...)
-	return nil
-}
-
-func (s *DependenciesStub) RunQuietlyAndIgnore(cmd string, stdin io.ReadSeeker, args ...string) {
-	s.execute(true /*quietly*/, cmd, args...)
-}
-
-func (s *DependenciesStub) execute(quietly bool, cmd string, args ...string) {
-	cmdline := strings.Join(append([]string{cmd}, args...), " ")
-	s.ExecutedAll = append(s.ExecutedAll, cmdline)
-	if quietly {
-		s.ExecutedQuietly = append(s.ExecutedQuietly, cmdline)
-	} else {
-		s.ExecutedNormally = append(s.ExecutedNormally, cmdline)
 	}
 }
