+++ tools/istio-clean-iptables/pkg/cmd/cleanup_test.go
@@ -88,20 +88,10 @@ func TestIptables(t *testing.T) {
 			cfg := constructTestConfig()
 			tt.config(cfg)
 
-			ext := &DependenciesStub{}
-			ext.DetectIptablesStub = dep.IptablesVersion{
-				DetectedBinary: "iptables",
-				DetectedSaveBinary: "iptables-save",
-				DetectedRestoreBinary: "iptables-restore",
-			}
-
-			ext.DetectIptablesV6Stub = dep.IptablesVersion{
-				DetectedBinary: "ip6tables",
-				DetectedSaveBinary: "ip6tables-save",
-				DetectedRestoreBinary: "ip6tables-restore",
-			}
-
-			cleaner := NewIptablesCleaner(cfg, &ext.DetectIptablesStub, &ext.DetectIptablesV6Stub, ext)
+			ext := &dep.DependenciesStub{}
+			iptStub, _ := ext.DetectIptablesVersion("", false)
+			ip6tStub, _ := ext.DetectIptablesVersion("", true)
+			cleaner := NewIptablesCleaner(cfg, &iptStub, &ip6tStub, ext)
 
 			cleaner.Run()
 
@@ -127,41 +117,4 @@ func compareToGolden(t *testing.T, name string, actual []string) {
 	gotBytes := []byte(strings.Join(actual, "\n"))
 	goldenFile := filepath.Join("testdata", name+".golden")
 	testutil.CompareContent(t, gotBytes, goldenFile)
-}
-
-// TODO BML replace DIY mocks/state with something better, also this is duplicated
-// with other stubs elsewhere
-type DependenciesStub struct {
-	ExecutedNormally []string
-	ExecutedQuietly  []string
-	ExecutedAll      []string
-	DetectIptablesStub dep.IptablesVersion
-	DetectIptablesV6Stub dep.IptablesVersion
-}
-
-func (s *DependenciesStub) Run(cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, stdin io.ReadSeeker, args ...string) error {
-	s.execute(false /*quietly*/, cmd, iptVer, args...)
-	return nil
-}
-
-func (s *DependenciesStub) RunQuietlyAndIgnore(cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, stdin io.ReadSeeker, args ...string) {
-	s.execute(true /*quietly*/, cmd, iptVer, args...)
-}
-
-// TODO BML this stub can be smarter
-func (s *DependenciesStub) DetectIptablesVersion(overrideVersion string, ipV6 bool) (dep.IptablesVersion, error) {
-	if ipV6 {
-		return s.DetectIptablesV6Stub, nil
-	}
-	return s.DetectIptablesStub, nil
-}
-
-func (s *DependenciesStub) execute(quietly bool, cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, args ...string) {
-	cmdline := strings.Join(append([]string{iptVer.CmdToString(cmd)}, args...), " ")
-	s.ExecutedAll = append(s.ExecutedAll, cmdline)
-	if quietly {
-		s.ExecutedQuietly = append(s.ExecutedQuietly, cmdline)
-	} else {
-		s.ExecutedNormally = append(s.ExecutedNormally, cmdline)
-	}
 }
