+++ tools/istio-iptables/pkg/capture/run_test.go
@@ -277,7 +277,7 @@ func TestIptables(t *testing.T) {
 			cfg := constructTestConfig()
 			tt.config(cfg)
 
-			ext := &DependenciesStub{}
+			ext := &dep.DependenciesStub{}
 			iptConfigurator := NewIptablesConfigurator(cfg, ext)
 			iptConfigurator.Run()
 			compareToGolden(t, tt.name, ext.ExecutedAll)
@@ -323,7 +323,7 @@ func TestSeparateV4V6(t *testing.T) {
 	for _, tt := range cases {
 		t.Run(tt.name, func(t *testing.T) {
 			cfg := constructTestConfig()
-			iptConfigurator := NewIptablesConfigurator(cfg, &dep.StdoutStubDependencies{})
+			iptConfigurator := NewIptablesConfigurator(cfg, &dep.DependenciesStub{})
 			v4Range, v6Range, err := iptConfigurator.separateV4V6(tt.cidr)
 			if err != nil {
 				t.Fatal(err)
@@ -343,48 +343,4 @@ func compareToGolden(t *testing.T, name string, actual []string) {
 	gotBytes := []byte(strings.Join(actual, "\n"))
 	goldenFile := filepath.Join("testdata", name+".golden")
 	testutil.CompareContent(t, gotBytes, goldenFile)
-}
-
-// TODO BML replace DIY mocks/state with something better, also this is duplicated
-// with other stubs elsewhere
-type DependenciesStub struct {
-	ExecutedNormally     []string
-	ExecutedQuietly      []string
-	ExecutedAll          []string
-}
-
-func (s *DependenciesStub) Run(cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, stdin io.ReadSeeker, args ...string) error {
-	s.execute(false /*quietly*/, cmd, iptVer, args...)
-	return nil
-}
-
-func (s *DependenciesStub) RunQuietlyAndIgnore(cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, stdin io.ReadSeeker, args ...string) {
-	s.execute(true /*quietly*/, cmd, iptVer, args...)
-}
-
-// TODO BML this stub can be smarter
-func (s *DependenciesStub) DetectIptablesVersion(overrideVersion string, ipV6 bool) (dep.IptablesVersion, error) {
-	if ipV6 {
-
-			return dep.IptablesVersion{
-				DetectedBinary:        "ip6tables",
-				DetectedSaveBinary:    "ip6tables-save",
-				DetectedRestoreBinary: "ip6tables-restore",
-			}, nil
-	}
-	return dep.IptablesVersion{
-		DetectedBinary:        "iptables",
-		DetectedSaveBinary:    "iptables-save",
-		DetectedRestoreBinary: "iptables-restore",
-	}, nil
-}
-
-func (s *DependenciesStub) execute(quietly bool, cmd constants.IptablesCmd, iptVer *dep.IptablesVersion, args ...string) {
-	cmdline := strings.Join(append([]string{iptVer.CmdToString(cmd)}, args...), " ")
-	s.ExecutedAll = append(s.ExecutedAll, cmdline)
-	if quietly {
-		s.ExecutedQuietly = append(s.ExecutedQuietly, cmdline)
-	} else {
-		s.ExecutedNormally = append(s.ExecutedNormally, cmdline)
-	}
 }
